Fix up delegate references in MainMenu.xib.
For the DragDropImageView class references in MainMenu.xib, once it was Swift instead of Objective-C, I *had* to specify the module or it wouldn't compile properly.

For the app delegate class, it doesn't appear it needed to be referenced in the MainMenu.xib at all to still work properly, and even when it was in the xib file, it didn't need to specify the module.

However, to be safe, I am specifying it in the xib file, and I am adding the module reference.